data_science/cloud_functions/TriggerProcessBattlesJob/requirements.txt
google-cloud-storage==1.25.0
google-api-python-client==1.7.11